refactor(strict): widen schema_version header field from u16 to u32
The binary tuple header stores a schema version so the decoder can
reconstruct the correct physical layout for multi-version reads.
Widening the field from u16 to u32 (4 bytes instead of 2) future-
proofs the format for schemas that undergo many ALTER cycles and
aligns the representation with the u32 version fields used in the
columnar and WAL formats.

Related `StrictSchema::version`, `ColumnDef::added_at_version`, and
`DroppedColumnDef::dropped_at_version` fields are widened to u32 in
`nodedb-types` to match.

+++ nodedb-types/src/columnar/column_type.rs
@@ -275,10 +275,10 @@ pub struct ColumnDef {
     /// the schema version after the bump so the reader can build a physical
     /// sub-schema for tuples written under older versions.
     #[serde(default = "default_added_at_version")]
-    pub added_at_version: u16,
+    pub added_at_version: u32,
 }
 
-fn default_added_at_version() -> u16 {
+fn default_added_at_version() -> u32 {
     1
 }
